Refactor Container block inspector controls
