+++ src/blocks/container/block.js
@@ -13,10 +13,12 @@ import { __ } from '@wordpress/i18n';
 import { registerBlockType } from '@wordpress/blocks';
 import { InnerBlocks } from '@wordpress/block-editor';
 import dynamicContentAttributes from '../../extend/dynamic-content/attributes';
+import { getBlockAttributes } from '../../block-context';
+import containerContext from '../../block-context/container';
 
 const attributes = Object.assign(
 	{},
-	blockAttributes,
+	getBlockAttributes( blockAttributes, containerContext, generateBlocksDefaults.container ),
 	dynamicContentAttributes
 );
