+++ src/extend/inspector-control/controls/background-panel/components/background-options.js
@@ -1,0 +1,65 @@
+import BackgroundColorOverlay from './background-color-overlay';
+import ImageSize from './image-size';
+import Selector from './selector';
+import ImageOpacity from './image-opacity';
+
+export default function BackgroundOptions( { attributes, setAttributes } ) {
+	const {
+		bgImage,
+		bgOptions,
+		bgImageSize,
+		useDynamicData,
+		dynamicContentType,
+		innerZindex,
+	} = attributes;
+
+	return (
+		<>
+			{ !! bgOptions.overlay ? (
+				<BackgroundColorOverlay
+					checked={ !! bgOptions.overlay }
+					onChange={ ( nextOverlay ) => {
+						setAttributes( {
+							bgOptions: {
+								...bgOptions,
+								overlay: nextOverlay,
+							},
+						} );
+					} }
+				/>
+			) : (
+				<>
+					{ ( ( bgImage && bgImage.id ) || ( useDynamicData && '' !== dynamicContentType ) ) &&
+						<ImageSize
+							value={ bgImageSize }
+							onChange={ ( value ) => setAttributes( { bgImageSize: value } ) }
+						/>
+					}
+
+					<Selector
+						value={ bgOptions.selector }
+						onChange={ ( value ) => {
+							setAttributes( { bgOptions: { ...bgOptions, selector: value } } );
+							if ( 'pseudo-element' === value && ! innerZindex && 0 !== innerZindex ) {
+								setAttributes( { innerZindex: 1 } );
+							}
+						} }
+					/>
+
+					<ImageOpacity
+						value={ bgOptions.opacity }
+						isPseudoElement={ 'pseudo-element' === bgOptions.selector }
+						onChange={ ( value ) => {
+							setAttributes( {
+								bgOptions: { ...bgOptions, opacity: value, selector: 'pseudo-element' },
+							} );
+							if ( ! innerZindex && 0 !== innerZindex ) {
+								setAttributes( { innerZindex: 1 } );
+							}
+						} }
+					/>
+				</>
+			) }
+		</>
+	);
+}
