+++ src/extend/inspector-control/controls/background-panel/components/position.js
@@ -1,0 +1,12 @@
+import { __ } from '@wordpress/i18n';
+import { TextControl } from '@wordpress/components';
+
+export default function Position( { value, onChange } ) {
+	return (
+		<TextControl
+			label={ __( 'Position', 'generateblocks' ) }
+			value={ value }
+			onChange={ onChange }
+		/>
+	);
+}
