+++ src/extend/inspector-control/controls/spacing/components/vertical-alignment.js
@@ -1,0 +1,14 @@
+import { __ } from '@wordpress/i18n';
+import { SelectControl } from '@wordpress/components';
+import spacingOptions from '../options';
+
+export default function VerticalAlignment( { value, onChange } ) {
+	return (
+		<SelectControl
+			label={ __( 'Vertical Alignment', 'generateblocks' ) }
+			value={ value }
+			options={ spacingOptions.verticalAlignment }
+			onChange={ onChange }
+		/>
+	);
+}
