+++ src/extend/inspector-control/controls/spacing/index.js
@@ -6,6 +6,9 @@ import { useDeviceType } from '../../../../hooks';
 import { useContext } from '@wordpress/element';
 import ControlsContext from '../../../../block-context';
 import DeviceControls from './components/device-controls';
+import MinimumHeight from './components/minimum-height';
+import VerticalAlignment from './components/vertical-alignment';
+import ZIndex from './components/z-index';
 
 export default function Spacing( { attributes, setAttributes, computedStyles } ) {
 	const [ device ] = useDeviceType();
@@ -20,6 +23,17 @@ export default function Spacing( { attributes, setAttributes, computedStyles } )
 		fillHorizontalSpace,
 		fillHorizontalSpaceTablet,
 		fillHorizontalSpaceMobile,
+		minHeight,
+		minHeightUnit,
+		verticalAlignment,
+		minHeightTablet,
+		minHeightUnitTablet,
+		verticalAlignmentTablet,
+		minHeightMobile,
+		minHeightUnitMobile,
+		verticalAlignmentMobile,
+		zindex,
+		innerZindex,
 	} = attributes;
 
 	return (
@@ -41,62 +55,134 @@ export default function Spacing( { attributes, setAttributes, computedStyles } )
 			}
 
 			{ 'Desktop' === device &&
-				<DeviceControls
-					inlineWidth={ !! inlineWidth }
-					onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidth: checked } ) }
-					stack={ !! stack }
-					onChangeStack={ ( value ) => {
-						setAttributes( {
-							stack: value,
-							stackTablet: !! value && ! stackTablet ? value : stackTablet,
-							stackMobile: !! value && ! stackMobile ? value : stackMobile,
-						} );
-					} }
-					fill={ !! fillHorizontalSpace }
-					onFillChange={ ( value ) => {
-						setAttributes( {
-							fillHorizontalSpace: value,
-							fillHorizontalSpaceTablet: !! value && ! fillHorizontalSpaceTablet ? value : fillHorizontalSpaceTablet,
-							fillHorizontalSpaceMobile: !! value && ! fillHorizontalSpaceMobile ? value : fillHorizontalSpaceMobile,
-						} );
-					} }
-				/>
+				<>
+					<DeviceControls
+						inlineWidth={ !! inlineWidth }
+						onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidth: checked } ) }
+						stack={ !! stack }
+						onChangeStack={ ( value ) => {
+							setAttributes( {
+								stack: value,
+								stackTablet: !! value && ! stackTablet ? value : stackTablet,
+								stackMobile: !! value && ! stackMobile ? value : stackMobile,
+							} );
+						} }
+						fill={ !! fillHorizontalSpace }
+						onFillChange={ ( value ) => {
+							setAttributes( {
+								fillHorizontalSpace: value,
+								fillHorizontalSpaceTablet: !! value && ! fillHorizontalSpaceTablet ? value : fillHorizontalSpaceTablet,
+								fillHorizontalSpaceMobile: !! value && ! fillHorizontalSpaceMobile ? value : fillHorizontalSpaceMobile,
+							} );
+						} }
+					/>
+
+					{ spacing.minimumHeight &&
+						<>
+							<MinimumHeight
+								value={ minHeight }
+								onChange={ ( value ) => setAttributes( { minHeight: parseFloat( value ) } ) }
+								unitValue={ minHeightUnit }
+								onChangeUnit={ ( value ) => setAttributes( { minHeightUnit: value } ) }
+							/>
+
+							{ !! minHeight && spacing.verticalAlignment &&
+								<VerticalAlignment
+									value={ verticalAlignment }
+									onChange={ ( value ) => setAttributes( { verticalAlignment: value } ) }
+								/>
+							}
+						</>
+					}
+
+					{ spacing.zIndex &&
+						<ZIndex
+							label={ spacing.innerZIndex && __( 'Outer z-index', 'generateblocks' ) }
+							value={ zindex }
+							onChange={ ( value ) => setAttributes( { zindex: value } ) }
+						/>
+					}
+
+					{ spacing.innerZIndex &&
+						<ZIndex
+							label={ __( 'Inner z-index', 'generateblocks' ) }
+							value={ innerZindex }
+							onChange={ ( value ) => setAttributes( { innerZindex: value } ) }
+						/>
+					}
+				</>
 			}
 
 			{ 'Tablet' === device &&
-				<DeviceControls
-					inlineWidth={ !! inlineWidthTablet }
-					onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidthTablet: checked } ) }
-					stack={ !! stackTablet }
-					onChangeStack={ ( value ) => {
-						setAttributes( {
-							stackTablet: value,
-							stackMobile: !! value && ! stackMobile ? value : stackMobile,
-						} );
-					} }
-					fill={ !! fillHorizontalSpaceTablet }
-					onFillChange={ ( value ) => {
-						setAttributes( {
-							fillHorizontalSpaceTablet: value,
-							fillHorizontalSpaceMobile: !! value && ! fillHorizontalSpaceMobile ? value : fillHorizontalSpaceMobile,
-						} );
-					} }
-				/>
+				<>
+					<DeviceControls
+						inlineWidth={ !! inlineWidthTablet }
+						onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidthTablet: checked } ) }
+						stack={ !! stackTablet }
+						onChangeStack={ ( value ) => {
+							setAttributes( {
+								stackTablet: value,
+								stackMobile: !! value && ! stackMobile ? value : stackMobile,
+							} );
+						} }
+						fill={ !! fillHorizontalSpaceTablet }
+						onFillChange={ ( value ) => {
+							setAttributes( {
+								fillHorizontalSpaceTablet: value,
+								fillHorizontalSpaceMobile: !! value && ! fillHorizontalSpaceMobile ? value : fillHorizontalSpaceMobile,
+							} );
+						} }
+					/>
+
+					{ spacing.minimumHeight &&
+						<>
+							<MinimumHeight
+								value={ minHeightTablet }
+								onChange={ ( value ) => setAttributes( { minHeightTablet: parseFloat( value ) } ) }
+								unitValue={ minHeightUnitTablet }
+								onChangeUnit={ ( value ) => setAttributes( { minHeightUnitTablet: value } ) }
+							/>
+
+							{ ( !! minHeight || !! minHeightTablet ) && spacing.verticalAlignment &&
+								<VerticalAlignment
+									value={ verticalAlignmentTablet }
+									onChange={ ( value ) => setAttributes( { verticalAlignmentTablet: value } ) }
+								/>
+							}
+						</>
+					}
+				</>
 			}
 
 			{ 'Mobile' === device &&
-				<DeviceControls
-					inlineWidth={ !! inlineWidthMobile }
-					onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidthMobile: checked } ) }
-					stack={ !! stackMobile }
-					onChangeStack={ ( value ) => {
-						setAttributes( { stackMobile: value } );
-					} }
-					fill={ !! fillHorizontalSpaceMobile }
-					onFillChange={ ( value ) => {
-						setAttributes( { fillHorizontalSpaceMobile: value } );
-					} }
-				/>
+				<>
+					<DeviceControls
+						inlineWidth={ !! inlineWidthMobile }
+						onChangeInlineWidth={ ( checked ) => setAttributes( { inlineWidthMobile: checked } ) }
+						stack={ !! stackMobile }
+						onChangeStack={ ( value ) => setAttributes( { stackMobile: value } ) }
+						fill={ !! fillHorizontalSpaceMobile }
+						onFillChange={ ( value ) => setAttributes( { fillHorizontalSpaceMobile: value } ) }
+					/>
+
+					{ spacing.minimumHeight &&
+						<>
+							<MinimumHeight
+								value={ minHeightMobile }
+								onChange={ ( value ) => setAttributes( { minHeightMobile: parseFloat( value ) } ) }
+								unitValue={ minHeightUnitMobile }
+								onChangeUnit={ ( value ) => setAttributes( { minHeightUnitMobile: value } ) }
+							/>
+
+							{ ( !! minHeight || !! minHeightTablet || !! minHeightMobile ) && spacing.verticalAlignment &&
+								<VerticalAlignment
+									value={ verticalAlignmentMobile }
+									onChange={ ( value ) => setAttributes( { verticalAlignmentMobile: value } ) }
+								/>
+							}
+						</>
+					}
+				</>
 			}
 		</PanelArea>
 	);
