+++ src/extend/inspector-control/index.js
@@ -10,8 +10,10 @@ import IconControls from './controls/icon';
 import ElementControls from './controls/element';
 import BackgroundGradient from './controls/background-gradient';
 import SettingsPanel from './controls/settings-panel';
+import BackgroundPanel from './controls/background-panel';
+import ShapesPanel from './controls/shapes-panel';
 
-export default function GenerateBlocksInspectorControls( { attributes, setAttributes, computedStyles, children } ) {
+export default function GenerateBlocksInspectorControls( { attributes, setAttributes, computedStyles = {}, children } ) {
 	const [ device ] = useDeviceType();
 	const {
 		supports: {
@@ -21,6 +23,8 @@ export default function GenerateBlocksInspectorControls( { attributes, setAttrib
 			spacing,
 			colors,
 			backgroundGradient,
+			backgroundPanel,
+			shapesPanel,
 			icon,
 			htmlTags,
 		},
@@ -66,6 +70,20 @@ export default function GenerateBlocksInspectorControls( { attributes, setAttrib
 
 			{ backgroundGradient.enabled &&
 				<BackgroundGradient
+					attributes={ attributes }
+					setAttributes={ setAttributes }
+				/>
+			}
+
+			{ backgroundPanel.enabled &&
+				<BackgroundPanel
+					attributes={ attributes }
+					setAttributes={ setAttributes }
+				/>
+			}
+
+			{ shapesPanel.enabled && 'Desktop' === device &&
+				<ShapesPanel
 					attributes={ attributes }
 					setAttributes={ setAttributes }
 				/>
